requirements.txt
# Core dependencies
anthropic>=0.39.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image handling
Pillow>=10.0.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Utilities
rich>=13.0.0
tqdm>=4.66.0
